+++ Cargo.toml
@@ -14,6 +14,7 @@ homepage = "https://github.com/rojo-rbx/tarmac"
 
 [dependencies]
 env_logger = "0.7.0"
+globset = "0.4.4"
 log = "0.4.8"
 reqwest = "0.9.20"
 serde = { version = "1.0", features = ["derive"] }
