+++ src/data/mod.rs
@@ -1,7 +1,5 @@
-mod input_config;
+mod config;
 mod manifest;
-mod project_config;
 
-pub use input_config::*;
+pub use config::*;
 pub use manifest::*;
-pub use project_config::*;
